+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - Migrate patient status view to Jetpack Compose
 - Migrate clinical decision support alert to Jetpack Compose
 - Bump sqlite-android version to 3.49.0
+- Replace `android-database-sqlcipher` library with `sqlcipher-android`
 
 ## 2025.05.20